[Automated Tests] Improve reporting and add expected failures
